+++ app/src/main/java/com/shamardn/podcasttime/ui/home/HomeInteractionListener.kt
@@ -1,5 +1,5 @@
 package com.shamardn.podcasttime.ui.home
 
 interface HomeInteractionListener {
-    fun onClickPodcast(trackId: Int)
+    fun onClickPodcast(trackId: Long)
 }
