+++ app/src/main/java/com/shamardn/podcasttime/ui/search/SearchInteractionListener.kt
@@ -1,5 +1,5 @@
 package com.shamardn.podcasttime.ui.search
 
 interface SearchInteractionListener {
-    fun onClickPodcast(trackId: Int)
+    fun onClickPodcast(trackId: Long)
 }
